feat: Allow overriding Foundry VTT data modules path via environment variable
This change introduces the `FOUNDTY_VTT_DATA_MODULES_PATH` environment
variable to allow users to specify a custom path for their Foundry VTT
data modules.

If the `FOUNDTY_VTT_DATA_MODULES_PATH` environment variable is set,
it will be used as the base path. Otherwise, the path will default
to the user's home directory as before. This provides greater
flexibility for development and deployment configurations.